Moved images to images folder
